requirements.txt
elevenlabs==2.14.0
fastapi==0.116.1
openai==1.106.1
pydantic==2.11.7
python-dotenv==1.1.1
uvicorn==0.35.0
